Move PagedList<T, TKey> to separate file
